Only one interaction type used in a given cog type
