+++ src/spellbot/cogs/ban_cog.py
