+++ src/spellbot/cogs/leave_cog.py
