+++ tests/cogs/test_block_cog.py
@@ -4,7 +4,7 @@ import pytest
 from discord_slash.context import MenuContext
 
 from spellbot import SpellBot
-from spellbot.cogs.block import BlockCog
+from spellbot.cogs.block_cog import BlockCog
 from spellbot.database import DatabaseSession
 from spellbot.models import Block, User
