+++ tests/cogs/test_events_cog.py
@@ -3,7 +3,7 @@ import pytest
 from discord_slash.context import InteractionContext
 
 from spellbot import SpellBot
-from spellbot.cogs.events import EventsCog
+from spellbot.cogs.events_cog import EventsCog
 from spellbot.database import DatabaseSession
 from spellbot.interactions import lfg_interaction
 from spellbot.models import Game, GameFormat, GameStatus, User
