+++ tests/cogs/test_leave_cog.py
@@ -1,6 +1,6 @@
 import pytest
 
-from spellbot.cogs.leave import LeaveGameCog
+from spellbot.cogs.leave_cog import LeaveGameCog
 from spellbot.interactions import leave_interaction
 from spellbot.models import Channel, Guild
 from tests.mixins import InteractionContextMixin
