+++ tests/cogs/test_lfg_cog.py
@@ -5,7 +5,7 @@ import pytest
 from sqlalchemy.sql.expression import update
 from syrupy.assertion import SnapshotAssertion
 
-from spellbot.cogs.lfg import LookingForGameCog
+from spellbot.cogs.lfg_cog import LookingForGameCog
 from spellbot.database import DatabaseSession
 from spellbot.interactions import leave_interaction, lfg_interaction
 from spellbot.models import Channel, Game, GameFormat, GameStatus, Guild, User, UserAward
