+++ tests/cogs/test_lfg_cog_points.py
@@ -4,7 +4,7 @@ import pytest
 from discord_slash.context import ComponentContext
 
 from spellbot import SpellBot
-from spellbot.cogs.lfg import LookingForGameCog
+from spellbot.cogs.lfg_cog import LookingForGameCog
 from spellbot.database import DatabaseSession
 from spellbot.models import GameStatus, Play
 from spellbot.settings import Settings
